+++ src/components/LoginForm/LoginForm.tsx
@@ -1,40 +1,10 @@
 import Link from 'components/@shared/Link';
-import PATH from 'constants/path';
-import { USER_MESSAGE } from 'constants/message';
-import authAPI from 'apis/auth';
-import { createInputValueGetter } from 'utils/dom';
 import styled from 'styled-components';
-import { useDispatch } from 'react-redux';
-import { useNavigate } from 'react-router-dom';
-import { userActions } from 'redux/actions';
 import LabeledInput from 'components/@shared/LabeledInput';
+import useLoginForm from './useLoginForm';
 
 function LoginForm() {
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (!(e.target instanceof HTMLFormElement)) return;
-
-    const formElement = e.target.elements;
-    const getInputValue = createInputValueGetter(formElement);
-    const user = {
-      username: getInputValue('id'),
-      password: getInputValue('password'),
-    };
-
-    try {
-      const userInfo = await authAPI.login(user);
-
-      dispatch(userActions.setUser(userInfo));
-      navigate(PATH.BASE);
-    } catch (error) {
-      if (error instanceof Error) {
-        alert(USER_MESSAGE.FAIL_LOGIN);
-      }
-    }
-  };
+  const { handleSubmit } = useLoginForm();
 
   return (
     <StyledForm onSubmit={handleSubmit}>
